backend/requirements.txt
flask
flask-cors
openai-whisper
transformers
torch
librosa
numpy
